Added store of withdraw

+++ migrations/20250519032653_create_withdraws_table.sql
@@ -1,7 +1,7 @@
 -- +goose Up
 -- +goose StatementBegin
 CREATE TABLE IF NOT EXISTS withdraws (
-    id bigint PRIMARY KEY NOT NULL,
+    id SERIAL PRIMARY KEY NOT NULL,
     user_id bigint NOT NULL,
     order_id bigint NOT NULL,
     points bigint DEFAULT 0,
